refactor(infra): move ica decoder logic
