+++ typescript/infra/src/tx/governance/types.ts
@@ -107,11 +107,6 @@ export interface GovernanceDecoderRuntime {
     chain: ChainName,
     tx: AnnotatedEV5Transaction,
   ): Promise<GovernTransaction>;
-  isIcaTransaction(chain: ChainName, tx: AnnotatedEV5Transaction): boolean;
-  readIcaTransaction(
-    chain: ChainName,
-    tx: AnnotatedEV5Transaction,
-  ): Promise<GovernTransaction>;
 }
 
 export interface DecodeContext {
